Add a datadog monitor plugin

+++ manager/monitor/mesos/mesos_test.go
@@ -50,10 +50,10 @@ func TestMesosMaster_GetUpdatedMetrics(t *testing.T) {
 	setupTest()
 	metrics, err := monitor.GetUpdatedMetrics([]string{"mesos.cluster.cpu_percent"})
 	assert.Nil(t, err)
-	assert.Equal(t, 25, (*metrics)[0].CurrentReading)
+	assert.Equal(t, float64(25), (*metrics)[0].CurrentReading)
 	metrics, err = monitor.GetUpdatedMetrics([]string{"mesos.cluster.mem_percent"})
 	assert.Nil(t, err)
-	assert.Equal(t, 50, (*metrics)[0].CurrentReading)
+	assert.Equal(t, float64(50), (*metrics)[0].CurrentReading)
 	metrics, err = monitor.GetUpdatedMetrics([]string{"invalid.metric.name"})
 	assert.NotNil(t, err)
 }
